feat: stability of prev tests + rm dpop
